+++ pyxem/tests/generators/test_subpixelrefinement_generator.py
@@ -19,15 +19,13 @@
 import pytest
 import numpy as np
 
-from pyxem.generators.subpixelrefinement_generator import (
-    SubpixelrefinementGenerator,
-    get_simulated_disc,
-    get_experimental_square,
-    get_simulated_disc,
-)
-from pyxem.signals.diffraction_vectors import DiffractionVectors
-from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from skimage import draw
+
+from pyxem.generators import SubpixelrefinementGenerator
+from pyxem.generators.subpixelrefinement_generator import (
+    get_simulated_disc, get_experimental_square
+)
+from pyxem.signals import DiffractionVectors, ElectronDiffraction2D
 
 
 @pytest.fixture()
@@ -51,12 +49,12 @@ def test_experimental_square_size(exp_disc):
 
 def test_failure_for_non_even_entry_to_get_simulated_disc():
     with pytest.raises(ValueError, match="'square_size' must be an even number"):
-        disc = get_simulated_disc(61, 5)
+        _ = get_simulated_disc(61, 5)
 
 
 def test_failure_for_non_even_errors_get_experimental_square(exp_disc):
     with pytest.raises(ValueError, match="'square_size' must be an even number"):
-        square = get_experimental_square(exp_disc, [17, 19], 7)
+        _ = get_experimental_square(exp_disc, [17, 19], 7)
 
 
 class Test_init_xfails:
@@ -70,7 +68,7 @@ class Test_init_xfails:
             ValueError,
             match="Some of your vectors do not lie within your diffraction pattern",
         ):
-            sprg = SubpixelrefinementGenerator(dp, vector)
+            _ = SubpixelrefinementGenerator(dp, vector)
 
     def test_out_of_range_vectors_DiffractionVectors(self):
         """Test that putting vectors that lie outside of the
@@ -82,7 +80,7 @@ class Test_init_xfails:
             ValueError,
             match="Some of your vectors do not lie within your diffraction pattern",
         ):
-            sprg = SubpixelrefinementGenerator(dp, vectors)
+            _ = SubpixelrefinementGenerator(dp, vectors)
 
     def test_wrong_navigation_dimensions(self):
         """Tests that navigation dimensions must be appropriate too."""
@@ -96,7 +94,7 @@ class Test_init_xfails:
             ValueError,
             match=r"Vectors with shape .* must have the same navigation shape as .*",
         ):
-            sprg = SubpixelrefinementGenerator(dp, vectors)
+            _ = SubpixelrefinementGenerator(dp, vectors)
 
 
 class set_up_for_subpixelpeakfinders: