+++ pyxem/tests/signals/test_diffraction_vectors.py
@@ -18,9 +18,11 @@
 
 import pytest
 import numpy as np
-from pyxem.signals.diffraction_vectors import DiffractionVectors
 from sklearn.cluster import DBSCAN
+
 from hyperspy.signals import Signal2D
+
+from pyxem.signals import DiffractionVectors
 
 # DiffractionVectors correspond to a single list of vectors, a map of vectors
 # all of equal length, and the ragged case. A fixture is defined for each of