+++ pyxem/tests/signals/test_lazy_diffraction1d.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyxem.signals.diffraction1d import Diffraction1D, LazyDiffraction1D
+from pyxem.signals import Diffraction1D, LazyDiffraction1D
 
 
 class TestDecomposition: